SDK-2245: net retrieve session
