+++ src/Yoti.Auth/DigitalIdentity/ShareSessionResult.cs
@@ -1,13 +1,12 @@
-﻿using System;
-using System.Reflection.Emit;
-using Newtonsoft.Json;
-using Yoti.Auth.DataObjects;
+﻿using Newtonsoft.Json;
+
 
 namespace Yoti.Auth.DigitalIdentity
 {
-    public class GetSessionResult
+    public class ShareSessionResult
     {
 #pragma warning disable 0649
+
         // These fields are assigned to by JSON deserialization
         [JsonProperty("id")]
         public string Id { get; set; }
@@ -18,9 +17,31 @@ namespace Yoti.Auth.DigitalIdentity
         [JsonProperty("expiry")]
         public string Expiry { get; set; }
 
+        [JsonProperty("created")]
+        public string Created { get; set; }
+
+        [JsonProperty("updated")]
+        public string Updated { get; set; }
+
+        [JsonProperty("qrCode")]
+        public qrCode QrCode { get; set; }
+
+        [JsonProperty("receipt")]
+        public receipt Receipt { get; set; }
+
 #pragma warning restore 0649
 
     }
 
-    
+    public class qrCode
+    {
+        [JsonProperty("id")]
+        public string Id { get; set; }
+    }
+
+    public class receipt
+    {
+        [JsonProperty("id")]
+        public string Id { get; set; }
+    }
 }
